fix: stop using sudo in recursive `make` invocations
We don't want to install with root if we're installing into a location
that doesn't require root access. For example, if we're creating a
binary release archive, we want to do that without sudo, or we'll be
giving users an archive with files that need root permissions.

Because of this, I had to get rid of the dependency on the `sand` and 
`sand-debug` targets for `install`. I never want to run `sudo make install`, 
have `make` notice the binary hasn't been build, and build it as root. Because 
then the next time I do `cargo build`, the cached artifacts are owned by root, 
so my unprivileged cargo just needs to discard them and do a full rebuild.
It's maddening.

The tension between these two issues honestly adds up to a really
annoying problem. I'm suprised I haven't seen more discussion of this
online, or that this isn't just a totally solved problem with some set
of best practices, or better tools. But `make` is ubiquitous.